+++ server/src/test/java/org/opensearch/ratelimitting/admissioncontrol/controllers/CPUBasedAdmissionControllerTests.java
@@ -11,7 +11,6 @@ package org.opensearch.ratelimitting.admissioncontrol.controllers;
 import org.opensearch.cluster.service.ClusterService;
 import org.opensearch.common.settings.ClusterSettings;
 import org.opensearch.common.settings.Settings;
-import org.opensearch.ratelimitting.admissioncontrol.AdmissionControlSettings;
 import org.opensearch.ratelimitting.admissioncontrol.enums.AdmissionControlMode;
 import org.opensearch.ratelimitting.admissioncontrol.settings.CPUBasedAdmissionControllerSettings;
 import org.opensearch.test.OpenSearchTestCase;
@@ -44,11 +43,11 @@ public class CPUBasedAdmissionControllerTests extends OpenSearchTestCase {
 
     public void testCheckDefaultParameters() {
         admissionController = new CPUBasedAdmissionController(
-            AdmissionControlSettings.CPU_BASED_ADMISSION_CONTROLLER,
+            CPUBasedAdmissionControllerSettings.CPU_BASED_ADMISSION_CONTROLLER,
             Settings.EMPTY,
             clusterService.getClusterSettings()
         );
-        assertEquals(admissionController.getName(), AdmissionControlSettings.CPU_BASED_ADMISSION_CONTROLLER);
+        assertEquals(admissionController.getName(), CPUBasedAdmissionControllerSettings.CPU_BASED_ADMISSION_CONTROLLER);
         assertEquals(admissionController.getRejectionCount(), 0);
         assertEquals(admissionController.settings.getTransportLayerAdmissionControllerMode(), AdmissionControlMode.DISABLED);
         assertFalse(admissionController.isEnabledForTransportLayer());
@@ -56,7 +55,7 @@ public class CPUBasedAdmissionControllerTests extends OpenSearchTestCase {
 
     public void testCheckUpdateSettings() {
         admissionController = new CPUBasedAdmissionController(
-            AdmissionControlSettings.CPU_BASED_ADMISSION_CONTROLLER,
+            CPUBasedAdmissionControllerSettings.CPU_BASED_ADMISSION_CONTROLLER,
             Settings.EMPTY,
             clusterService.getClusterSettings()
         );
@@ -68,7 +67,7 @@ public class CPUBasedAdmissionControllerTests extends OpenSearchTestCase {
             .build();
         clusterService.getClusterSettings().applySettings(settings);
 
-        assertEquals(admissionController.getName(), AdmissionControlSettings.CPU_BASED_ADMISSION_CONTROLLER);
+        assertEquals(admissionController.getName(), CPUBasedAdmissionControllerSettings.CPU_BASED_ADMISSION_CONTROLLER);
         assertEquals(admissionController.getRejectionCount(), 0);
         assertEquals(admissionController.settings.getTransportLayerAdmissionControllerMode(), AdmissionControlMode.ENFORCED);
         assertTrue(admissionController.isEnabledForTransportLayer());
@@ -76,14 +75,14 @@ public class CPUBasedAdmissionControllerTests extends OpenSearchTestCase {
 
     public void testApplyControllerWithDefaultSettings() {
         admissionController = new CPUBasedAdmissionController(
-            AdmissionControlSettings.CPU_BASED_ADMISSION_CONTROLLER,
+            CPUBasedAdmissionControllerSettings.CPU_BASED_ADMISSION_CONTROLLER,
             Settings.EMPTY,
             clusterService.getClusterSettings()
         );
         assertEquals(admissionController.getRejectionCount(), 0);
         assertEquals(admissionController.settings.getTransportLayerAdmissionControllerMode(), AdmissionControlMode.DISABLED);
         action = "indices:data/write/bulk[s][p]";
-        admissionController.accept(action);
+        admissionController.apply(action);
         assertEquals(admissionController.getRejectionCount(), 0);
     }
 
@@ -95,33 +94,14 @@ public class CPUBasedAdmissionControllerTests extends OpenSearchTestCase {
             )
             .build();
         admissionController = new CPUBasedAdmissionController(
-            AdmissionControlSettings.CPU_BASED_ADMISSION_CONTROLLER,
+            CPUBasedAdmissionControllerSettings.CPU_BASED_ADMISSION_CONTROLLER,
             settings,
             clusterService.getClusterSettings()
         );
         assertTrue(admissionController.isEnabledForTransportLayer());
         assertEquals(admissionController.getRejectionCount(), 0);
         action = "indices:data/write/bulk[s][p]";
-        admissionController.accept(action);
+        admissionController.apply(action);
         assertEquals(admissionController.getRejectionCount(), 1);
-    }
-
-    public void testApplyControllerWhenSettingsEnabledWithInvalidAction() {
-        Settings settings = Settings.builder()
-            .put(
-                CPUBasedAdmissionControllerSettings.CPU_BASED_ADMISSION_CONTROLLER_TRANSPORT_LAYER_MODE.getKey(),
-                AdmissionControlMode.ENFORCED.getMode()
-            )
-            .build();
-        admissionController = new CPUBasedAdmissionController(
-            AdmissionControlSettings.CPU_BASED_ADMISSION_CONTROLLER,
-            settings,
-            clusterService.getClusterSettings()
-        );
-        assertTrue(admissionController.isEnabledForTransportLayer());
-        assertEquals(admissionController.getRejectionCount(), 0);
-        action = "TEST";
-        admissionController.accept(action);
-        assertEquals(admissionController.getRejectionCount(), 0);
     }
 }